Implemented INotifyPropertyChanged Interfaces in MainWindow and Options Window View Models

+++ ViewModel/VMMainWindow.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -10,16 +11,18 @@ using System.Windows.Input;
 
 namespace ViewModel
 {
-    public class VMMainWindow
+    public class VMMainWindow : INotifyPropertyChanged
     {
         private ObservableCollection<VMHabiticaTodo> todoList;
-        
+
         private String errorMsg;
         private HabiticaClient client;
         private ICommand fetchCommand;
         private ICommand sendCommand;
         private ICommand createCommand;
         private ICommand deleteCommand;
+        public event PropertyChangedEventHandler PropertyChanged;
+
         public VMMainWindow()
         {
             TodoList = new ObservableCollection<VMHabiticaTodo>();
@@ -27,16 +30,23 @@ namespace ViewModel
             CreateCommand = new UserCommand(new Action<object>(CreateNewTodo));
             DeleteCommand = new UserCommand(new Action<object>(DeleteTodo));
             SendCommand = new UserCommand(new Action<object>(SendTodos));
+            ErrorMsg = "";
             GetHabiticaClientInstance();
         }
+        public void OnPropertyChanged(PropertyChangedEventArgs e)
+        {
+            if (PropertyChanged != null)
+                PropertyChanged(this, e);
+        }
+
 
         private void GetHabiticaClientInstance()
         {
-            ErrorMsg = "";
+            
             try
             {
                 client = HabiticaClient.GetInstance();
-                
+
             }
             catch (NoCredentialsException e)
             {
@@ -47,72 +57,89 @@ namespace ViewModel
         private async void SendTodos(object obj)
         {
             GetHabiticaClientInstance();
-            try
+            if (client != null)
             {
-                await client.SaveTodo(((VMHabiticaTodo)obj).Todo);
+                try
+                {
+                    await client.SaveTodo(((VMHabiticaTodo)obj).Todo);
+                }
+                catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
+                {
+                    ErrorMsg = e.Message;
+                }
             }
-            catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
-            {
-                ErrorMsg = e.Message;
-            }
-            
 
         }
 
         private async void DeleteTodo(object obj)
         {
             GetHabiticaClientInstance();
-            try
+            if (client != null)
             {
-                await client.DeleteTodo(((VMHabiticaTodo)obj).Todo);
+                try
+                {
+                    await client.DeleteTodo(((VMHabiticaTodo)obj).Todo);
+                }
+                catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
+                {
+                    ErrorMsg = e.Message;
+                }
             }
-            catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
-            {
-                ErrorMsg = e.Message;
-            }
-
         }
 
         private async void CreateNewTodo(object obj)
         {
             GetHabiticaClientInstance();
-            try
+            if (client != null)
             {
-                TodoList.Add(new VMHabiticaTodo(await client.CreateNewTodo("new Todo")));
+                try
+                {
+                    TodoList.Add(new VMHabiticaTodo(await client.CreateNewTodo("new Todo")));
+                }
+                catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
+                {
+                    ErrorMsg = e.Message;
+                }
             }
-            catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
-            {
-                ErrorMsg = e.Message;
-            }
-
         }
 
-        
+
         private async void FetchTodos(object o)
         {
             GetHabiticaClientInstance();
             TodoList.Clear();
-            try
-            {
-                IList<HabiticaTodo> templist = await client.GetTodos();
-                foreach (HabiticaTodo h in templist)
+            if(client!=null)
+            {                
+                try
                 {
-                    todoList.Add(new VMHabiticaTodo(h));
+                    IList<HabiticaTodo> templist = await client.GetTodos();
+                    foreach (HabiticaTodo h in templist)
+                    {
+                        todoList.Add(new VMHabiticaTodo(h));
+                    }
                 }
-            }
-            catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
-            {
-                ErrorMsg = e.Message;
+                catch (Exception e) when (e is WrongCredentialsException || e is UnsuccessfulException)
+                {
+                    ErrorMsg = e.Message;
+                }
             }
 
         }
         public ObservableCollection<VMHabiticaTodo> TodoList { get => todoList; set => todoList = value; }
         public ICommand FetchCommand { get => fetchCommand; set => fetchCommand = value; }
-       
+
 
         public ICommand CreateCommand { get => createCommand; set => createCommand = value; }
         public ICommand DeleteCommand { get => deleteCommand; set => deleteCommand = value; }
         public ICommand SendCommand { get => sendCommand; set => sendCommand = value; }
-        public string ErrorMsg { get => errorMsg; set => errorMsg = value; }
+        public string ErrorMsg
+        {
+            get => errorMsg;
+            set
+            {
+                errorMsg = value;
+                OnPropertyChanged(new PropertyChangedEventArgs("ErrorMsg"));
+            }
+        }
     }
 }
